refactor: move event ticket confirmation details into hooks
