+++ src/EventTickets/Actions/AddEventTicketsToDonationConfirmationPageDonationTotal.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Give\EventTickets\Actions;
+
+use Give\EventTickets\Models\EventTicket;
+use Give\EventTickets\Repositories\EventTicketRepository;
+use Give\Framework\Receipts\DonationReceipt;
+use Give\Framework\Receipts\Properties\ReceiptDetail;
+use Give\Framework\Support\ValueObjects\Money;
+
+/**
+ * @unreleased
+ */
+class AddEventTicketsToDonationConfirmationPageDonationTotal
+{
+    /**
+     * @unreleased
+     */
+    public function __invoke(DonationReceipt $receipt): void
+    {
+        $eventTickets = give(EventTicketRepository::class)->queryByDonationId($receipt->donation->id)->getAll();
+
+        if (!empty($eventTickets)) {
+            $currency = $receipt->donation->amount->getCurrency();
+            $total = array_reduce($eventTickets, function (Money $carry, EventTicket $eventTicket) {
+                $ticketType = $eventTicket->ticketType()->get();
+
+                return $carry->add(
+                    $ticketType->price
+                );
+            }, new Money(0, $currency));
+
+            $receipt->donationDetails->addDetail(
+                new ReceiptDetail(
+                    __('Event Tickets', 'give'),
+                    $total->formatToLocale()
+                )
+            );
+        }
+    }
+}
